Extract codec to module
